requirements.txt
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
certifi==2026.2.25
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
executing==2.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.13.0
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
openai==2.24.0
packaging==26.0
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.2
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.47
stack-data==0.6.3
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
wcwidth==0.6.0
